Moved fault tests to own class
